Updates and bug fixes to the stable branch

+++ Makefile
@@ -7,7 +7,8 @@ PWD := $(shell echo $$PWD)
 USER_ID := $(shell id -u)
 GROUP_ID := $(shell id -g)
 UNAME_S := $(shell uname -s)
-S3_BUCKET := safe-client-libs-jenkins
+S3_BUCKET := safe-jenkins-build-artifacts
+UUID := $(shell uuidgen | sed 's/-//g')
 
 build-container:
 	rm -rf target/
@@ -87,6 +88,22 @@ package-deploy-artifacts:
 		-u ${USER_ID}:${GROUP_ID} \
 		maidsafe/safe-client-libs-build:${SAFE_APP_VERSION} \
 		scripts/package-runner-container
+
+retrieve-cache:
+ifndef SCL_BRANCH
+	@echo "A branch reference must be provided."
+	@echo "Please set SCL_BRANCH to a valid branch reference."
+	@exit 1
+endif
+ifeq ($(OS),Windows_NT)
+	aws s3 cp \
+		--no-sign-request \
+		--region eu-west-2 \
+		s3://${S3_BUCKET}/scl-${SCL_BRANCH}-windows-cache.tar.gz .
+endif
+	mkdir target
+	tar -C target -xvf scl-${SCL_BRANCH}-windows-cache.tar.gz
+	rm scl-${SCL_BRANCH}-windows-cache.tar.gz
 
 retrieve-build-artifacts:
 ifndef SCL_BRANCH
@@ -171,26 +188,6 @@ endif
 	rm ${SCL_BRANCH}-${SCL_BUILD_NUMBER}-scl-osx-x86_64.tar.gz
 	rm ${SCL_BRANCH}-${SCL_BUILD_NUMBER}-scl-mock-osx-x86_64.tar.gz
 
-test-artifacts-mock:
-ifeq ($(UNAME_S),Linux)
-	docker run --rm -v "${PWD}":/usr/src/safe_client_libs:Z \
-		-u ${USER_ID}:${GROUP_ID} \
-		-e CARGO_TARGET_DIR=/target \
-		-e SCL_TEST_SUITE=mock \
-		maidsafe/safe-client-libs-build:${SAFE_APP_VERSION} \
-		scripts/test-runner-container
-else
-	./scripts/test-mock
-endif
-
-test-artifacts-integration:
-	docker run --rm -v "${PWD}":/usr/src/safe_client_libs:Z \
-		-u ${USER_ID}:${GROUP_ID} \
-		-e CARGO_TARGET_DIR=/target \
-		-e SCL_TEST_SUITE=integration \
-		maidsafe/safe-client-libs-build:${SAFE_APP_VERSION} \
-		scripts/test-runner-container
-
 test-artifacts-binary:
 ifndef SCL_BCT_PATH
 	@echo "A value must be supplied for the previous binary compatibility test suite."
@@ -209,15 +206,30 @@ endif
 		scripts/test-runner-container
 
 tests: clean
+	rm -rf artifacts
+ifeq ($(UNAME_S),Linux)
 	rm -rf target/
-	docker run --name safe_app_build \
+	docker run --name "safe_app_tests-${UUID}" \
 		-v "${PWD}":/usr/src/safe_client_libs \
 		-u ${USER_ID}:${GROUP_ID} \
 		-e CARGO_TARGET_DIR=/target \
 		maidsafe/safe-client-libs-build:${SAFE_APP_VERSION} \
 		scripts/test-mock
-	docker cp safe_app_build:/target .
-	docker rm -f safe_app_build
+	docker cp "safe_app_tests-${UUID}":/target .
+	docker rm -f "safe_app_tests-${UUID}"
+else
+	./scripts/test-mock
+endif
+	mkdir artifacts
+	find target/release -maxdepth 1 -type f -exec cp '{}' artifacts \;
+
+tests-integration: clean
+	rm -rf target/
+	docker run --rm -v "${PWD}":/usr/src/safe_client_libs \
+		-u ${USER_ID}:${GROUP_ID} \
+		-e CARGO_TARGET_DIR=/target \
+		maidsafe/safe-client-libs-build:${SAFE_APP_VERSION} \
+		scripts/test-integration
 
 debug:
 	docker run --rm -v "${PWD}":/usr/src/crust maidsafe/safe-client-libs-build:${SAFE_APP_VERSION} /bin/bash
